Modified the update manager
now game objects that needs to be updated need to use the Updatable script, that adds and removes itself from the update manager when being enablaed or disabled.

that way the update manager doent iterate through disabaled Iupdatables. by doing so, I had to change the foreach to a for.